pyproject.toml
[tool.poetry]
name = "vuln-rag-bot"
version = "0.1.0"
description = "MVP for a Vulnerability Analysis RAG bot"
authors = ["Your Team"]
packages = [{ include = "app", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pydantic-ai-slim = {version = "^1.0", extras = ["openai"]}
openai = "^1.0"
fastapi = "^0.111"
uvicorn = "^0.29"
asyncpg = "^0.29"
psycopg2-binary = "^2.9"
mitreattack-python = "^2.0"
markdown-it-py = "^3.0"
pydantic-settings = "^2.2"
click = "^8.1"
setuptools = "^70.0"
beautifulsoup4 = "^4.12"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
ruff = "^0.4"
mypy = "^1.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true